refactor: moving the frontdoor helpers into the frontdoor package
