+++ azurerm/internal/services/frontdoor/frontdoor_custom_https_configuration_resource.go
@@ -54,7 +54,7 @@ func resourceArmFrontDoorCustomHttpsConfiguration() *schema.Resource {
 				Optional: true,
 				MaxItems: 1,
 				Elem: &schema.Resource{
-					Schema: azure.SchemaFrontdoorCustomHttpsConfiguration(),
+					Schema: SchemaFrontdoorCustomHttpsConfiguration(),
 				},
 			},
 		},
@@ -141,7 +141,7 @@ func resourceArmFrontDoorCustomHttpsConfigurationRead(d *schema.ResourceData, me
 
 	if resp.Name != nil {
 
-		flattenedHttpsConfig := azure.FlattenArmFrontDoorCustomHttpsConfiguration(*resp.FrontendEndpointProperties)
+		flattenedHttpsConfig := FlattenArmFrontDoorCustomHttpsConfiguration(*resp.FrontendEndpointProperties)
 		if err := d.Set("custom_https_configuration", flattenedHttpsConfig.CustomHTTPSConfiguration); err != nil {
 			return fmt.Errorf("setting `custom_https_configuration`: %+v", err)
 		}
